schedule_backend/requirements.txt
Django
djangorestframework
Pillow
psycopg2
pytz
uwsgi
requests
drf-yasg
